Fixed issue where when pressing keys rapidly while moving player, player might get kind of stuck and not move fluently. Did this by adding a boolean array in the KeyInput class to keep track of when exactly keys are being pressed down.  Created new enemy type SmartEnemyHad to change everything except rectangles to floats to get new smart enemy to work properly.

+++ LilJavaGame/src/com/liljavagame/main/HUD.java
@@ -4,16 +4,16 @@ import java.awt.*;
 
 public class HUD {
 
-    public static int HEALTH = 100;
+    public static float HEALTH = 100;
 
-    private int greenValue = 255;
+    private float greenValue = 255f;
 
     private int score = 0;
     private int level = 1;
 
     public void tick() {
-        HEALTH = Game.clamp(HEALTH, 0, 100); // Ensures health will not run outside(left or right) of health bar
-        greenValue = Game.clamp(greenValue, 0, 255);
+        HEALTH = (int) Game.clamp(HEALTH, 0, 100); // Ensures health will not run outside(left or right) of health bar
+        greenValue = (int) Game.clamp(greenValue, 0, 255);
 
         greenValue = HEALTH * 2;
 
@@ -23,8 +23,8 @@ public class HUD {
     public void render(Graphics g) {
         g.setColor(Color.gray);
         g.fillRect(15, 15, 200, 32);
-        g.setColor(new Color(75, greenValue, 0));
-        g.fillRect(15, 15, HEALTH * 2, 32);
+        g.setColor(new Color(75, (int) greenValue, 0));
+        g.fillRect(15, 15, (int) (HEALTH * 2), 32);
         g.setColor(Color.white);
         g.drawRect(15, 15, 200, 32); // puts border on health bar
